Installable units holded by P2Repository are now sorted in descending order
The main consequence is that `P2Repository.findIU` now returns a *sorted*
sequence of `InstallableUnit`, where the most recent one comes first.
Furthermore, `P2Repository.findIU` now expect an exact `unitId` to be passed,
and no longer works with regular expressions (the functionality was removed
because it is not needed).

Also, since we need to compare OSGI Version numbers, I've added a dependency to
"org.osgi" % "org.osgi.core" % "4.3.0", as it provides a `Version` data object
that implementes `Comparable`. This lead to a few changes so use `Version`
instead of a raw string.