New --force=<ecosystem-ids> CLI argument to force recreation of ecosystem update-sites
The existing implementation of `EcosystemBuild.shouldBeRegenerated`, which takes care of
deciding if an ecosystem update-site should be recreated, doesn't currently handle the
case where a plug-in is removed from the ecosystem and we would like to recreate the
ecosystem update-site **without** the removed plug-in. The information about what plug-in
was removed wrt the previously built update-site is missing from `ScalaIDEVersion` class.

After discussing the problem with @skyluc, it was decided to implement a hack to override
the behavior of `EcosystemBuild.shouldBeRegenerated` and manually force the recreation of
the update-site via a CLI argument `--force=<comma-separated-list-of-ecosystem-ids>`. The
rationale for this decision is that a proper fix would need some work (not something that
can be quickly put together), and considering that we don't remove plug-ins from the
ecosystem often, we decided it was better to have a workaround than further delay the removal
of InSynth from the ecosystem.